requirements.txt
flask-rethinkdb
flask-security
feedparser
apscheduler
listparser
beautifulsoup4
opml
requests
git+git://github.com/rgv151/rwrapper.git
validate_email
xxhash